fix: PR Requests (3/4 done)

+++ server/src/main/kotlin/delta/codecharacter/server/daily_challenge/DailyChallengeSeeder.kt
@@ -2,34 +2,42 @@ package delta.codecharacter.server.daily_challenge
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.context.event.ApplicationReadyEvent
 import org.springframework.context.event.EventListener
 import org.springframework.stereotype.Component
-import java.io.File
+
+
 
 @Component
 class DailyChallengeSeeder {
 
     @Autowired private lateinit var dailyChallengeRepository: DailyChallengeRepository
 
-    @Value(
-        "/server/server/src/main/kotlin/delta/codecharacter/server/daily_challenge/dcConstants.json"
-    )
-    lateinit var filePath: String
+    private val logger : Logger = LoggerFactory.getLogger(DailyChallengeSeeder::class.java)
     @EventListener(ApplicationReadyEvent::class)
     fun doSomethingAfterStartup() {
-        println("Initialising daily challenges")
 
-        //      Delete everything initially
-        dailyChallengeRepository.deleteAll()
+        if(dailyChallengeRepository.findAll().isEmpty()){
+            logger.info("Seeding daily_challenges")
 
-        //      Seed the database
-        val jsonString = File(filePath).readText()
-        val objectMapper = jacksonObjectMapper()
-        val dcs: List<DailyChallengeEntity> = objectMapper.readValue(jsonString)
+            //      Delete everything initially
+            dailyChallengeRepository.deleteAll()
 
-        dailyChallengeRepository.saveAll(dcs)
+            //      Seed the database
+            val jsonString = this::class.java.classLoader.getResource("dcConstants.json").readText()
+            if(jsonString != null){
+                val objectMapper = jacksonObjectMapper()
+                val dcs: List<DailyChallengeEntity> = objectMapper.readValue(jsonString)
+                dailyChallengeRepository.saveAll(dcs)
+            }
+            else{
+                logger.error("dcConstants.json not found or doesn't exist")
+            }
+
+        }
+
     }
 }
